requirements.txt
# SAMA 2.0 - Successifier Autonomous Marketing Agent
# Python 3.12+ required

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & Agent Framework
anthropic==0.18.1
langgraph==0.0.26
langchain==0.1.6
langchain-anthropic==0.1.1

# Workflow Engine
temporal-sdk==1.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4
alembic==1.13.1

# Vector Store
pinecone-client==3.0.2

# Cache & Queue
redis==5.0.1
celery==5.3.6

# Google APIs
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# SEO & Marketing APIs
semrush-api==0.1.0  # Note: May need custom implementation
ahrefs-api==0.1.0   # Note: May need custom implementation

# Social Media
tweepy==4.14.0

# HTTP & Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Monitoring & Logging
sentry-sdk==1.40.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
pytz==2024.1
croniter==2.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx-mock==0.1.0

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0
